style: 코드 및 파일 정리
