--- CloudNotes/CloudNotes/Strings.swift
@@ -1,14 +1,0 @@
-import Foundation
-
-enum UserDefaultsKeys: String {
-    case selectedMemoIndexPathRow = "selectedMemoIndexPathRow"
-    case isCellSelected = "isCellSelected"
-}
-
-enum NotificationName: String {
-    case updateTableViewList = "updateTableViewList"
-    case deleteCell = "deleteCell"
-    case moveCellToTop = "moveCellToTop"
-}
-
-
