--- CloudNotes/CloudNotes/Supporting Files/CloudNotes.xcdatamodeld/CloudNotes.xcdatamodel/contents
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="17709" systemVersion="19H2" minimumToolsVersion="Automatic" sourceLanguage="Swift" usedWithCloudKit="YES" userDefinedModelVersionIdentifier="">
-    <entity name="Memo" representedClassName="MemoMO" syncable="YES" codeGenerationType="class">
-        <attribute name="body" optional="YES" attributeType="String"/>
-        <attribute name="lastModified" optional="YES" attributeType="Date" usesScalarValueType="NO"/>
-        <attribute name="title" optional="YES" attributeType="String"/>
-    </entity>
-    <elements>
-        <element name="Memo" positionX="-36" positionY="9" width="128" height="88"/>
-    </elements>
-</model>
